+++ src/messages/en/admin.php
@@ -3,7 +3,7 @@
 return [
     'language_name' => 'English',
     'button_save' => 'Save',
-    'button_confirm' => 'Yes',
+    'button_yes' => 'Yes',
     'button_abort' => 'Cancel',
     'button_send' => 'Send',
     'login_pre_title' => 'Login for {title}',
@@ -352,4 +352,16 @@ return [
     'view_index_redirect_telephone' => 'telephone number',
     'view_index_redirect_telephone_help' => 'Please enter a valid telephone number (for "+" sign a leading "\" is require).',
     'missing_file_upload_permission' => 'Insufficient permissions for file uploads.',
+    
+// 1.2.2
+    
+    'button_no' => 'No',
+    'crud_exportdata_btn' => 'Export Data',
+    'crud_exportdata_col_header' => 'Header',
+    'crud_exportdata_col_format' => 'Format',
+    'crud_exportdata_col_format_csv' => 'CSV',
+    'crud_exportdata_col_format_xlsx' => 'Excel (xlsx)',
+    'crud_exportdata_col_columns' => 'Columns',
+    'crud_exportdata_btn_generateexport' => 'Generate export',
+    'crud_exportdata_btn_downloadexport' => 'Download export',
 ];