+++ src/messages/ua/admin.php
@@ -3,7 +3,7 @@
 return [
     'language_name' => 'Українська',
     'button_save' => 'Зберегти',
-    'button_confirm' => 'Так',
+    'button_yes' => 'Так',
     'button_abort' => 'Відміна',
     'button_send' => 'Відправити',
     'login_pre_title' => 'Вхід для {title}',
@@ -353,4 +353,16 @@ return [
     'view_index_redirect_telephone' => 'telephone number',
     'view_index_redirect_telephone_help' => 'Please enter a valid telephone number (for "+" sign a leading "\" is require).',
     'missing_file_upload_permission' => 'Insufficient permissions for file uploads.',
+    
+// 1.2.2
+    
+    'button_no' => 'No',
+    'crud_exportdata_btn' => 'Export Data',
+    'crud_exportdata_col_header' => 'Header',
+    'crud_exportdata_col_format' => 'Format',
+    'crud_exportdata_col_format_csv' => 'CSV',
+    'crud_exportdata_col_format_xlsx' => 'Excel (xlsx)',
+    'crud_exportdata_col_columns' => 'Columns',
+    'crud_exportdata_btn_generateexport' => 'Generate export',
+    'crud_exportdata_btn_downloadexport' => 'Download export',
 ];