+++ src/ngrest/render/RenderCrud.php
@@ -73,14 +73,16 @@ class RenderCrud extends Render implements ViewContextInterface, RenderCrudInter
     }
     
     /**
-     * 
-     * @return NULL[]
+     * Generates an array with all attributes an the corresponding label.
+     *
+     * @return array
+     * @since 1.2.2
      */
     public function generateDownloadAttributes()
     {
         $attributes = [];
-        foreach ($this->model->attributes as $key => $atr) {
-            $attributes[$key] = $this->model->getAttributeLabel($key);
+        foreach ($this->model->attributes() as $key) {
+            $attributes[$key] = $this->model->getAttributeLabel($key) . ' ('.$key.')';
         }
         
         return $attributes;